\name{createThreeWayDataset}
\alias{colorblind}
\alias{blankplot}
\alias{pal}
\alias{statemaps}

\title{
colorblind
}
\description{
colorblind
}
\usage{
colorblind (a, lo=min(a,na.rm=TRUE), hi=max(a,na.rm=TRUE),
        n.shades=1001)

}
\arguments{
\item{a}{}
\item{lo}{}
\item{hi}{}
\item{n.shades}{}
}
\details{
colorblind
}
\source{

}
\references{

}
\examples{
}
\keyword{manip}
