\name{createThreeWayDataset}
\alias{createThreeWayDataset}

\title{
createThreeWayDataset
}
\description{
createThreeWayDataset
}
\usage{
createThreeWayDataset(y, var1, var2, var3, 
  pop.weight=rep(1, length(y)), 
  positiveResponse=levels(y)[2])

}
\arguments{
\item{y}{}
\item{var1}{}
\item{var2}{}
\item{var3}{}
\item{pop.weight}{}
\item{positiveResponse}{}
}
\details{
createThreeWayDataset
}
\source{

}
\references{

}
\examples{
}
\keyword{manip}
