\name{getNumberWays-methods}
\docType{methods}
\alias{getNumberWays-methods}
\alias{getNumberWays,mrp-method}
\title{Get the number of ways for the analysis}
\description{Returns the depth of the analysis. This number is generated from the data set that is used to generate the \code{\link{mrp}} 
    object.}
\keyword{methods}