\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,NWayData-method}
\alias{getData,mrp-method}
\alias{getData-method}
\alias{getEstimates}
\alias{getEstimates,mrp-method}
\alias{getEstimates-method}
\alias{getFormula}
\alias{getFormula,mrp-method}
\alias{getFormula-method}
\alias{getModel}
\alias{getModel,mrp-method}
\alias{getModel-method}
\alias{getOutcome}
\alias{getOutcome,mrp-method}
\alias{getOutcome-method}
\alias{getPopulation}
\alias{getPopulation,mrp-method}
\alias{getPopulation-method}
\alias{getResponse}
\alias{getResponse,mrp-method}
\alias{getResponse-method}
\alias{getThetaHat}
\alias{getThetaHat,mrp-method}
\alias{getThetaHat-method}
\alias{setFormula}
\alias{setFormula,mrp-method}
\alias{setFormula-method}
\alias{setPopulation}
\alias{setPopulation,mrp-method}
\alias{setPopulation-method}
\alias{setPopulationOnes}
\alias{setPopulationOnes,mrp-method}
\alias{setPopulationOnes-method}
\alias{show}
\alias{show,NWayData-method}
\alias{show,mrp-method}
\title{mrp object methods}
\description{
  Mostly getters and setters for its slots
}

