\name{MRP-package}
\Rdversion{1.1}
\alias{MRP-package}
\alias{MRP}
\docType{package}
\title{
  Multilevel Regression and Poststratification (MRP) Package
}
\description{
  This package contains a user-friendly implementation of multilevel regression and poststratification.
}
\details{
  This package also includes implementations of basic display functionality for three-way analysis. 
  These visualizers are designed to be extended for more complex display visual analysis.
  For a complete list of functions, use library(help="MRP").  
}

}

\author{
  Andrew Gelman <gelman@stat.columbia.edu>,
  Daniel Lee <bearlee@alum.mit.edu>,
  Yu-Sung Su <ys463@columbia.edu>,

  Maintainer: Daniel Lee <bearlee@alum.mit.edu>
}
% TODO: Fill in references section
%\references{
%
%}
\keyword{ package }
s