\name{mrp.theme}
\alias{mrp.theme}
\title{Generate a lattice theme with top and left strip titles}
\usage{
  mrp.theme(rowlength, collength)
}
\arguments{
  \item{rowlength}{numeric, number of rows in a matrix of
  maps}

  \item{collength}{numeric, number of columns in a matrix
  of maps}
}
\value{
  a list suitable for \code{par.settings} or
  \code{trellis.par.set}
}
\description{
  Generate a lattice theme with top and left strip titles
}

