\name{serialPaste}
\alias{serialPaste}
\title{serial paste}
\usage{
  serialPaste(x, collapse = "and")
}
\arguments{
  \item{x}{vector or list}

  \item{collapse}{default="and"}
}
\description{
  Function to paste together a list of items, separated by
  commas (if more than 2), and with the last one having the
  collapse string.
}

