\name{mrp-class}
\Rdversion{1.1}
\docType{class}
\alias{mrp-class}
\alias{mrp}

\title{Multilevel regression and poststratification object}
\description{
	The \code{mrp} class is the core class of the MRP package. A \code{mrp}
	object is typically created with the data of interest and the object is
	then used to perform multilevel regression and poststratification.
}
\section{Objects from the Class}{
	Objects can be created by calls of the form \code{newMrp (response, var1, var2, var3, weight, positiveResponse, formula)}. 
	The only required elements of the constructor are \code{response}, \code{var1}, \code{var2}, and \code{var3}. 
	This will set up the \code{mrp} object for logistic regression. 
	 ~~ describe objects here ~~ 
}

\section{Slots}{
	\describe{
    	\item{\code{data}:}{The raw input data (class \code{"data.frame"}). The \code{data} is typically initialized on construction.}
    	\item{\code{numberWays}:}{The depth of the analysis (type \code{"integer"}). Note: for now, this is fixed at 3. As this package matures, we will allow for more complex analyses.}
    	\item{\code{data.nWay}:}{The transformed data (class \code{"ThreeWayData"}). The depth of the data must match the variable in slot \code{numberWays}.}
    	\item{\code{formula}:}{The formula for the multilevel regression (class \code{"character"}, e.g. a string representation).}
    	\item{\code{multilevelModel}:}{The multilevel regression model (class \code{"mer"} created by \code{"glmer"})}
    	\item{\code{theta.hat}:}{The estimates generated by the multilevel regression (class \code{"array"}). This can be indexed by the indexes of \code{data}.}
    	\item{\code{population}:}{The population distribution (class \code{"array"}). This is intended to be a probability mass table (all entries sum to 1), 
    			but the package will also work with unnormalized population distributions. The effect may be present in the visualization step.}
    	\item{\code{.debug}:}{This variable was meant to be a way to provide additional information (Object of class \code{"logical"}).}
  }
}
\section{Methods}{
	setPopulation, 
	setFormula, 
	getFormula, 
	mr, 
	p, 
	createNWayData, 
	hasNWayData, 
	fitMultilevelModel, 
	hasMultilevelModel
}
\references{ ~put references to the literature/web site here ~ }
\author{Andrew Gelman <gelman@stat.columbia.edu>,
	Daniel Lee <bearlee@alum.mit.edu>,
	Yu-Sung Su <ys463@columbia.edu>}
\note{ ~~further notes~~ }

 ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
	~~objects to See Also as \code{\link{~~fun~~}}, ~~~
	or \code{\linkS4class{CLASSNAME}} for links to other classes
}
\examples{
%showClass("mrp")
}
\keyword{classes}
