\name{MRP-package}
\Rdversion{1.1}
\alias{MRP-package}
\alias{MRP}
\docType{package}
\title{
	Multilevel Regression and Poststratification
}
\description{
	A user-friendly implementation of multilevel regression and poststratification. 
	This package includes some basic display functionality for three-way analysis.
	The visualizers are designed to be extended for extended display functionality. 
}
\details{
\tabular{ll}{
	Package: \tab MRP\cr
	Version: \tab 0.00.01\cr
	Date: \tab 2009-9-21\cr
	Depends: \tab methods, arm, lme4, maps, colorspace\cr
	License: \tab GPL (>=2)\cr
	Collate: \tab 'classThreeWayData.R' 'classMrp.R' 'classMrpVisualizer.R' 'classMrpStateVisualizer.R'\cr
	Packaged: \tab 2009-12-08 20:05:30 UTC; Statistics\cr
	Built: \tab R 2.9.1; ; 2009-12-08 20:05:33 UTC; windows\cr
}

Index:
\preformatted{
	colorblind              colorblind
	createThreeWayDataset   createThreeWayDataset
	visualizeThreewayDataset
                        visualizeThreewayDataset
}

~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
	Andrew Gelman <gelman@stat.columbia.edu>,
	Daniel Lee <bearlee@alum.mit.edu>,
	Yu-Sung Su <ys463@columbia.edu>,

	Maintainer: Daniel Lee <bearlee@alum.mit.edu>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
